Started implementing authentication and authorization

+++ app/Http/Controllers/HairSalonController.php
@@ -2,43 +2,47 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\City;
 use App\Models\HairSalon;
 use Illuminate\Http\Request;
-use App\Http\Requests\HairSalonRequest;
+use Illuminate\Database\QueryException;
 use App\Http\Resources\HairSalonResource;
-use App\Filters\HairSalonFilter;
+use App\Http\Requests\StoreHairSalonRequest;
+use App\Http\Requests\UpdateHairSalonRequest;
 
 class HairSalonController extends Controller
 {
-    public function index(Request $request)
+    public function index(City $city)
     {
-        $filter = new HairSalonFilter();
-        $queries = $filter->transform($request);
-        if(count($queries) != 0)
-        {
-            return response(HairSalonResource::collection(HairSalon::where($queries)->get()), 200);
-        }
-        return response(HairSalonResource::collection(HairSalon::all()), 200);
+        return response(HairSalonResource::collection($city->HairSalons), 200);
     }
 
-    public function show(HairSalon $salon)
+    public function show(City $city, HairSalon $salon)
     {
+        if($city->id != $salon->city_id) return response(['message' => 'resource not found'], 404);
         return response(new HairSalonResource($salon), 200);
     }
 
-    public function store(HairSalonRequest $request)
+    public function store(City $city, StoreHairSalonRequest $request)
     {
+        $payload = auth()->payload();
+        $request->merge([
+            'city_id' => $city->id,
+            'manager_id' => (int)$payload['sub']
+        ]);
         return response(new HairSalonResource(HairSalon::create($request->all())), 201);
     }
 
-    public function update(HairSalon $salon, HairSalonRequest $request)
+    public function update(City $city, HairSalon $salon, UpdateHairSalonRequest $request)
     {
+        if($city->id != $salon->city_id) return response(['message' => 'resource not found'], 404);
         $salon->update($request->all());
         return response(new HairSalonResource($salon),200);
     }
 
-    public function delete(HairSalon $salon)
+    public function delete(City $city, HairSalon $salon)
     {
+        if($city->id != $salon->city_id) return response(['message' => 'resource not found'], 404);
         $salon->delete();
         return response('', 204);
     }
